Fixed type conversion warnings in unit tests

+++ tests/EnumFlags.cpp
@@ -23,7 +23,7 @@ TEST_CASE("EnumFlags")
 	CHECK(!flags.Has(Values::Three));
 	CHECK(!flags.Has(Values::Four));
 	CHECK(flags.Value() == 0);
-	CHECK(flags == 0b0000);
+	CHECK(flags == static_cast<uint16_t>(0b0000));
 
 	SECTION("OR-ing Flags")
 	{
@@ -33,7 +33,7 @@ TEST_CASE("EnumFlags")
 		CHECK(!flags.Has(Values::Three));
 		CHECK(!flags.Has(Values::Four));
 		CHECK(flags.Value() == 0b0001);
-		CHECK(flags == 0b0001);
+		CHECK(flags == static_cast<uint16_t>(0b0001));
 
 		flags |= EnumFlags{Values::Two};
 		CHECK(flags.Has(Values::One));
@@ -41,7 +41,7 @@ TEST_CASE("EnumFlags")
 		CHECK(!flags.Has(Values::Three));
 		CHECK(!flags.Has(Values::Four));
 		CHECK(flags.Value() == 0b0011);
-		CHECK(flags == 0b0011);
+		CHECK(flags == static_cast<uint16_t>(0b0011));
 
 		flags |= 0b0100;
 		CHECK(flags.Has(Values::One));
@@ -49,7 +49,7 @@ TEST_CASE("EnumFlags")
 		CHECK(flags.Has(Values::Three));
 		CHECK(!flags.Has(Values::Four));
 		CHECK(flags.Value() == 0b0111);
-		CHECK(flags == 0b0111);
+		CHECK(flags == static_cast<uint16_t>(0b0111));
 
 		flags |= Values::Four;
 		CHECK(flags.Has(0b0001));
@@ -57,7 +57,7 @@ TEST_CASE("EnumFlags")
 		CHECK(flags.Has(0b0100));
 		CHECK(flags.Has(0b1000));
 		CHECK(flags.Value() == 0b1111);
-		CHECK(flags == 0b1111);
+		CHECK(flags == static_cast<uint16_t>(0b1111));
 	}
 
 	SECTION("AND-ing Flags")
@@ -68,7 +68,7 @@ TEST_CASE("EnumFlags")
 		CHECK(flags.Has(Values::Three));
 		CHECK(flags.Has(Values::Four));
 		CHECK(flags.Value() == 0b1100);
-		CHECK(flags == 0b1100);
+		CHECK(flags == static_cast<uint16_t>(0b1100));
 
 		SECTION("Test 1")
 		{
@@ -78,7 +78,7 @@ TEST_CASE("EnumFlags")
 			CHECK(flags.Has(Values::Three));
 			CHECK(!flags.Has(Values::Four));
 			CHECK(flags.Value() == 0b0100);
-			CHECK(flags == 0b0100);
+			CHECK(flags == static_cast<uint16_t>(0b0100));
 		}
 
 		SECTION("Test 2")
@@ -89,7 +89,7 @@ TEST_CASE("EnumFlags")
 			CHECK(flags.Has(Values::Three));
 			CHECK(flags.Has(Values::Four));
 			CHECK(flags.Value() == 0b1100);
-			CHECK(flags == 0b1100);
+			CHECK(flags == static_cast<uint16_t>(0b1100));
 		}
 
 		SECTION("Test 3")
@@ -100,7 +100,7 @@ TEST_CASE("EnumFlags")
 			CHECK(!flags.Has(Values::Three));
 			CHECK(flags.Has(Values::Four));
 			CHECK(flags.Value() == 0b1000);
-			CHECK(flags == 0b1000);
+			CHECK(flags == static_cast<uint16_t>(0b1000));
 		}
 	}
 
@@ -110,5 +110,5 @@ TEST_CASE("EnumFlags")
 	CHECK(!flags.Has(Values::Three));
 	CHECK(!flags.Has(Values::Four));
 	CHECK(flags.Value() == 0b0000);
-	CHECK(flags == 0b0000);
+	CHECK(flags == static_cast<uint16_t>(0b0000));
 }
